requirements.txt
streamlit
transformers
openai
pyyaml